Grafana dashboards upload (#257)
Ok, so we decided to create grafana dashboards from the server instead of relying on the default by-file implementation. We are creating those the first time an application is created, that is, when we first report data for an application. That's the first moment we know the platform of the app reporting. We are using Grafana's HTTP API in order to create dashboards but first, we have to make sure there are no dashboards in Grafana already (either db dashboards or file dashboards). Once created, we have to change the home dashboard to one of the ones we have uploaded, identified by its slug: `db/home` (if we are still working with a file dashboard we won't do anything).