iOS UI and Disk dashboards (#259)
* Add the new UI and Disk dashboards, they are meant to be uploaded just as usual with the system designed in #257. 
* Add the new field to store user defaults size in bytes. Keep in mind that we are not sending the view controller being shown when the FPS are calculated so I've removed the possibility to group reports by screen.
* Remove check preventing server to upload dashboards if there were already dashboards in grafana.
* Add script to update dashboards in grafana programmatically.
We are using the grafana HTTP API just as we are doing in the GrafanaClient class in server. The script just iterates all over a range of organization ids, upload the dashboards and then finds the one named db/home, iterate all over the organization users and set the home dashboard as that one.
The script uses jq to parse JSON from console, if you don't have it installed then you better do before running anything.
The script is a one-time script actually and it shouldn't be used anymore but I want to upload it just in case we need it for any reason.