iOS configuration page (#269)
Create another page explaining how to configure the iOS SDK, very similar to the current Android getting started page. The version is retrieved from cocoapods API which will be the only way to distribute the SDK for now. Instead of using a single Getting Started link a list of two elements, Android and iOS is displayed, redirecting the user to the proper page.

Clean up, including:
* Rename some classes and methods to stop being Android-centric.
* Fix some typos, spacings and styles.