New billing system (#277)
Start using Stripe in replacement of Braintree, the main reason being that Stripe lets us do what we want, that is, charge a dynamic price for a subscription (depending on the number of devices the client wants to report).

We are still using Taxamo to apply taxes correctly, however we will be using the Stripe JS client and do the linking of the charges between the two services in a different way as described here: https://www.taxamo.com/doc/payment_providers/stripe/subscription_api/

To summarize the article: we have to first create a placeholder transaction in Taxamo with some custom fields so that it knows what charge does it have to apply taxes to. Then, we need to create the Stripe subscription with a custom field containing the Taxamo transaction key. In both services we had to configure several integrations and webhooks.