Active FlowUp users email retrieval script (#279)
Add a SQL query that takes all the organizations that have an app with the maximum number of devices already reporting us data. From those orgs, we take all their users and extract their emails and names.

Fix issue when generating the javadoc for the project and update the CI configuration to make sure there are no builds with broken javadoc being merged.