Separate applications by platform (#282)
The right way to implement this is to add another field to the applications table and make it part of the key but that's a long process right now so we are doing the short version, that is, adding a suffix to iOS app package names. By doing so, two reports coming from the same app name but from different platforms are going to be separated into two different applications.

Show in the sidebar the platform of the applications, if they don't have any suffix then we are going to expect it to be from Android and iOS apps will need the suffix in order to be identified as so.

The suffixing solution is being done in the ApplicationRepository through a normalization step and showing the platform of the application in the sidebar is done through a view model + mapper that is responsible for showing the right platform depending on if the iOS suffix is present or not.